Added support for collectionfactory spawning in dmsdk.
Replicated the already existing style and functionality of comp_factory.h but for collectionfactories.
