+++ engine/gamesys/src/gamesys/components/comp_collection_factory.cpp
@@ -407,5 +407,13 @@ namespace dmGameSystem
         component->m_CustomResource = resource;
     }
 
+    bool CompCollectionFactorySpawn(HCollectionFactoryComponent component, dmGameObject::HCollection collection,
+                                                const dmVMath::Point3& position, const dmVMath::Quat& rotation, const dmVMath::Vector3& scale,
+                                                dmGameObject::InstancePropertyBuffers* properties, dmGameObject::InstanceIdMap* out_instances)
+    {
+        return dmGameObject::SpawnFromCollection(collection, CompCollectionFactoryGetResource(component)->m_CollectionDesc, properties,
+                                                         position, rotation, scale, out_instances);
+    }
+
 }
