+++ engine/gamesys/src/gamesys/components/comp_collection_factory.h
@@ -16,51 +16,27 @@
 #define DM_GAMESYS_COLLECTION_FACTORY_H
 
 #include <gameobject/component.h>
-#include <resource/resource.h>
-
+//#include <resource/resource.h>
+#include <dmsdk/gamesys/components/comp_collection_factory.h>
 namespace dmGameSystem
 {
-    struct CollectionFactoryComponent;  // for scripting
-    struct CollectionFactoryWorld;      // for scripting
-    struct CollectionFactoryResource;   // for scripting
-
     dmGameObject::CreateResult CompCollectionFactoryNewWorld(const dmGameObject::ComponentNewWorldParams& params);
-
     dmGameObject::CreateResult CompCollectionFactoryDeleteWorld(const dmGameObject::ComponentDeleteWorldParams& params);
-
     dmGameObject::CreateResult CompCollectionFactoryCreate(const dmGameObject::ComponentCreateParams& params);
-
     dmGameObject::CreateResult CompCollectionFactoryDestroy(const dmGameObject::ComponentDestroyParams& params);
-
     dmGameObject::CreateResult CompCollectionFactoryAddToUpdate(const dmGameObject::ComponentAddToUpdateParams& params);
-
     dmGameObject::UpdateResult CompCollectionFactoryUpdate(const dmGameObject::ComponentsUpdateParams& params, dmGameObject::ComponentsUpdateResult& update_result);
-
     dmGameObject::PropertyResult CompCollectionFactoryGetProperty(const dmGameObject::ComponentGetPropertyParams& params, dmGameObject::PropertyDesc& out_value);
 
-    // For scripting
+    dmResource::HFactory        CompCollectionFactoryGetResourceFactory(CollectionFactoryWorld* world);
 
     bool CompCollectionFactoryLoad(CollectionFactoryWorld* world, CollectionFactoryComponent* component, int callback_ref, int self_ref, int url_ref);
     bool CompCollectionFactoryUnload(CollectionFactoryWorld* world, CollectionFactoryComponent* component);
 
-    /**
-     * CompCollectionFactoryStatus
-     */
-    enum CompCollectionFactoryStatus
-    {
-        COMP_COLLECTION_FACTORY_STATUS_UNLOADED = 0,//!< COMP_COLLECTION_FACTORY_STATUS_UNLOADED
-        COMP_COLLECTION_FACTORY_STATUS_LOADING = 1, //!< COMP_COLLECTION_FACTORY_STATUS_LOADING
-        COMP_COLLECTION_FACTORY_STATUS_LOADED = 2,  //!< COMP_COLLECTION_FACTORY_STATUS_LOADED
-    };
-
-    dmResource::HFactory        CompCollectionFactoryGetResourceFactory(CollectionFactoryWorld* world);
-    CompCollectionFactoryStatus CompCollectionFactoryGetStatus(CollectionFactoryComponent* component);
-    bool                        CompCollectionFactoryIsDynamicPrototype(CollectionFactoryComponent* component);
-    bool                        CompCollectionFactoryIsLoading(CollectionFactoryComponent* component);
+    void                        CompCollectionFactorySetResource(CollectionFactoryComponent* component, CollectionFactoryResource* resource);
     CollectionFactoryResource*  CompCollectionFactoryGetResource(CollectionFactoryComponent* component);
     CollectionFactoryResource*  CompCollectionFactoryGetDefaultResource(CollectionFactoryComponent* component);
     CollectionFactoryResource*  CompCollectionFactoryGetCustomResource(CollectionFactoryComponent* component);
-    void                        CompCollectionFactorySetResource(CollectionFactoryComponent* component, CollectionFactoryResource* resource);
 }
 
 #endif
